requirements.txt
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
pytest-asyncio==0.21.1
pytest-cov==4.1.0 